utils: Remove 'splash-remote' tool.
This was a tool which aimed to use a reverse SSH tunnel to
provide access to a running detector, mainly for access while deployed
during the Splash Adelaide Festival. It didn't really work as reliably as
expected.

This functionalty has been replaced with a 'zerotier' network,
and setup and configured elsewhere.