Adds NodeRED configuration for saving events to a file
This work is the start of a project to migrate the existing
cosmic-array software functionality using the NodeRED framework.

The committed file holds the definition of a flow which takes the
events from seen on the GPIO pins and logs this to a file with a
timestamp.

To this example on a cosmic-array detector with NodeRED installed,
copy the flow json definition file to the ~/.node-red/lib/flows
directory, and import the configuration via the web interface.

The events will be save in the ~/events.log file.